+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - Add validation of picodata.yaml config
+- Add `--picodata-path` option to `config apply` command
 - Add optional `--no-build` flag to `plugin pack` command
 
 ### Changed